--- packages/web/src/hooks/usePullRequests.ts
@@ -1,11 +1,0 @@
-import { useAtomValue } from 'jotai/utils';
-import { pullRequestsReducerAtom } from '../jotai/pullRequest';
-
-export const usePullRequests = () => {
-  const { pullRequests, firstLoading } = useAtomValue(pullRequestsReducerAtom);
-
-  return {
-    firstLoading,
-    pullRequests,
-  };
-};
