+++ packages/web/src/stores/index.ts
@@ -1,2 +1,3 @@
 export { useSetting } from './setting/useSetting';
 export { useAuth } from './auth/useAuth';
+export { usePullRequests, useWatchPullRequests } from './pullRequest';
