+++ packages/web/src/stores/pullRequest/index.ts
@@ -1,0 +1,2 @@
+export { usePullRequests } from './usePullRequests';
+export { useWatchPullRequests } from './useWatchPullRequests';
