+++ packages/web/test/mocks/factory/index.ts
@@ -1,0 +1,4 @@
+export * from './pullRequest';
+export * from './repository';
+export * from './setting';
+export * from './user';
